Hamil: added preprocessor #if statements which     remove all Windows related code in the project     when !defined(_MSVC_VER) (temporary!),   switched the compiler to clang,   fixed (some - most serious) compiler warnings,   the __PROJECT_DIR #define is now created via CMake's     configure_file() Hamil/extern/sysv: added unix static/shared library files     needed to compile the project Hamil/build.eugene: made the input folder for (reqourcegen)     a variable util/structpack.h: added macros which allow cross-platform     packing/removing padding of structures Eugene: improvements to cross-platform compatibility,   fixed bugs Eugene/eugene_sysv.c: eugene_sysv_GetDateTimeFormat now returns     a properly formatted date string Eugene/extern/sysv: added unix variant of libclang.dll

+++ Hamil/src/mesh/util.cpp
@@ -15,9 +15,9 @@ std::tuple<std::vector<PNVertex>, std::vector<u16>> sphere(uint rings, uint sect
   verts.reserve(rings*sectors);
   for(uint r = 0; r < rings; r++) {
     for(uint s = 0; s < sectors; s++) {
-      auto y = sin(-(PIf/2.0f) + PIf*r*R);
-      auto x = cos(2.0f*PIf * s * S) * sin(PIf * r * R);
-      auto z = sin(2.0f*PIf * s * S) * sin(PIf * r * R);
+      auto y = sinf(-(PIf/2.0f) + PIf*r*R);
+      auto x = cosf(2.0f*PIf * s * S) * sinf(PIf * r * R);
+      auto z = sinf(2.0f*PIf * s * S) * sinf(PIf * r * R);
 
       // Flush very small values to 0.0f
       if(fabs(x) < 1e-6f) x = 0.0f;
